FIX TypeScript Example: Babel exclude undefined (#112)
* TypeScript Example: Fix Babel Exclude

Since the format was changed, the exclude is undefined from the rules array. Leading babel to crash by processing `node_modules` and probably the rest of the internet if we let it... 😄

* Set sourceMap for Style-Loader Instances
Post-Css Loader only has one instance in the webpack build process if it is used multiple times. The options are only set once. To be able to use post-css to its fullest, enable sourceMaps so path-throughs work properly.

* Fix Linting of webpack.js (no-unused-var)