🚀 7.2.0 maintenance (#1293)
* Proper typings

* Don't run eslint for type definitions

* Make less strict and auto infer

* Add back Routes renderFn

* Remove console.log

* Switch to data-prefetch

The prefetch attribute is not an attribute on anchor elements. The preferred way to do custom attributes is using data-* attributes or custom web elements.

When using anchor attributes the only potential candidate would be rel="prefetch", but that is not enabled for anchor elements (only link elements):

https://developer.mozilla.org/en-US/docs/Web/HTTP/Link_prefetching_FAQ

* Format

* Fix inverted option

* Format

* Bump versions within major range

* Add missing configuration

This is a dependency from eslint-config-react-tools but not listed in the dependencies of that package.

* Use browsers configuration from project

This removes the hardcoded browserlists by injecting them from the root path.

This is a "sorta breaking" change, as it changes the defaults from:
- >1%,
- last 4 versions,
- Firefox ESR,
- not ie < 9

To:

- > 0.5%
- last 2 versions
- Firefox ESR
- not dead

Since this is also a bug, it will be released with the minor release. Adding `browserslist` to your package.json or a config file, instantly fixes this.

#closes #1126

* Formatting

* Remove hardcoded fix

The fix works, which means that the browserlist functionality works. It now pulls it automaticaly from the project (user project) config, or defaults to "defaults".

* Bump packages and add devPeers

It's common to add peerDependencies as devDependencies. This way, they _will_ be installed on development and thus you can test as if they're really there.

Some of the peerDependencies were missing in the first place.

Additionally, during development, we want to use the local react-static folder module, which is symlinked.

* Remove mapped names, update mocks and fix tests

Now that the configuration returns absolute paths, this updates the tests to also "mock" absolute paths.

* Update template versions

* Update urls to use react-static

Nozzle is no longer owning this repository.

* Update CHANGELOG and add RELEASE

* v7.2.0

* Collapse sites list